Take NoopMessageHandler out of ITCursorResetToLatestTest so that it can be reused
